Added Search functions to BlackSheep. Beware of changes made to Sol13D. I was debugging BS comparing it to Sol1. Might have introduced an error

+++ Sol13D/Search.cpp
@@ -8,7 +8,7 @@
 #include <cfloat>
 
 const int Search::sMaxIter = 500000;
-const int Search::sNominalValue = 1.0;
+const double Search::sNominalValue = 1.0;
 const double Search::SEARCH_INF = DBL_MAX;
 
 /**
